+++ libfmt/runevsmprint.c
@@ -30,7 +30,7 @@ runeFmtStrFlush(Fmt *f)
 
 	if(f->start == nil)
 		return 0;
-	n = (int)f->farg;
+	n = (uintptr_t)f->farg;
 	n *= 2;
 	s = (Rune*)f->start;
 	f->start = realloc(s, sizeof(Rune)*n);
@@ -41,7 +41,7 @@ runeFmtStrFlush(Fmt *f)
 		free(s);
 		return 0;
 	}
-	f->farg = (void*)n;
+	f->farg = (void*)(uintptr_t)n;
 	f->to = (Rune*)f->start + ((Rune*)f->to - s);
 	f->stop = (Rune*)f->start + n - 1;
 	return 1;
@@ -61,7 +61,7 @@ runefmtstrinit(Fmt *f)
 	f->to = f->start;
 	f->stop = (Rune*)f->start + n - 1;
 	f->flush = runeFmtStrFlush;
-	f->farg = (void*)n;
+	f->farg = (void*)(uintptr_t)n;
 	f->nfmt = 0;
 	return 0;
 }
